The Freestyle Club/requirements.txt
discord.py==1.5.0
asyncpg
aioredis